+++ R/emis_grid.R
@@ -10,17 +10,17 @@
 #' Also, it is required that, when is a SpatialLinesDataFrame, there is a field
 #' called lkm, with the length of the road, in this case, in km.
 #' This function accepts data with "units" but they are converted internally
-#' to numeric and then return SpatialPolygonsDataFrame with numeric data.frame
+#' to numeric and then return "sf" with numeric data.frame.
 #'
-#' @param spobj A spatial dataframe of class sp
-#' @param g A grid with class SpatialPolygonsDataFrame
-#' @param sr Spatial reference, default is "+init=epsg:4326"
-#' @param type type of geometry: "lines" or "points"
-#' @importFrom raster intersect
-#' @importFrom rgeos gLength
-#' @importFrom stats aggregate
-#' @import sp
-#' @import rgdal
+#' @param spobj A spatial dataframe of class "sp" or "sf". When class is "sp"
+#' it is transformed to "sf".
+#' @param g A grid with class "SpatialPolygonsDataFrame" or "sf".
+#' @param sr Spatial reference. It must be for projected data, e.g: 31983.
+#' @param type type of geometry: "lines" or "points".
+#' @param array Logical to return GriddedEmissionsArray or Polygons for sf.
+#' @importFrom sf st_sf st_dimension st_transform st_length st_cast st_intersection
+#' @importFrom sp over
+#' @importFrom data.table data.table
 #' @export
 #' @note A future version of VEIN will imports or depend on the new package
 #' 'spatial features'. The migration for VEIN will be converting the 'Spatial'
@@ -55,38 +55,61 @@
 #' E_CO_STREETS <- emis_post(arra = E_CO, pollutant = "CO", by = "streets_wide")
 #' net@data <- cbind(net@data, E_CO_STREETS)
 #' head(net@data)
-#' g <- make_grid(net, 1/102.47/2, 1/102.47/2, polygon = T) #500m in degrees
+#' g <- make_grid(net, 1/102.47/2, 1/102.47/2) #500m in degrees
 #' net@data <- net@data[,- c(1:9)]
 #' names(net)
-#' E_CO_g <- emis_grid(spobj = net, g = g, sr= "+init=epsg:31983")
-#' head(E_CO_g@data)
+#' E_CO_g <- emis_grid(spobj = net, g = g, sr= 31983)
+#' head(E_CO_g) #class sf
+#' E_CO_g$V138 <- as.numeric(E_CO_g$V138)
+#' E_CO_g <- as(E_CO_g, "Spatial")
 #' spplot(E_CO_g, "V138", scales=list(draw=T),cuts=8,
 #' colorkey = list(space = "bottom", height = 1),
 #' col.regions = rev(bpy.colors(9)),
 #' sp.layout = list("sp.lines", net, pch = 16, cex = 2, col = "black"))
 #' }
-emis_grid <- function(spobj, g, sr, type="lines"){
-  if (type == "lines" ) {
-    net <- spobj
-    for( i in 1:ncol(net@data) ){
-      net@data[,i] <- as.numeric(net@data[,i])
-    }
-    net$lkm <-  rgeos::gLength(sp::spTransform(net,CRS(sr)),byid = T)/1000
-    netg <- raster::intersect(net,g)
-    netg$lkm2 <-  rgeos::gLength(sp::spTransform(netg,CRS(sr)),byid = T)/1000
-    netg@data[,1:(ncol(netg@data)-3)] <-  netg@data[,1:(ncol(netg@data)-3)] * netg$lkm2/netg$lkm
-    dfm <- stats::aggregate(cbind(netg@data[,1:(ncol(netg@data)-3)]),
-                            by=list(netg$id), sum, na.rm=TRUE)
-    colnames(dfm)[1] <- "id"
-    gg <- merge(g, dfm, by="id")
-    # for(i in 2:ncol(gg)){
-    #   gg@data[,i] <- gg@data[,i] * units::as_units("g h-1")
-    # } # spplot does not work with units
-    return(gg)
-  } else if ( type == "points" ){
-      g@data <- sp::over(g,spobj, fn=sum)
-      #Add units
-      return(g)
-    }
+emis_grid <- function(spobj, g, sr, type = "lines"){
+  net <- spobj
+  if(class(net) == "SpatialLinesDataFrame"){
+    net <- sf::st_as_sf(net)
+    net$id <- NULL
+  } else if(class(g) == "SpatialPolygonsDataFrame"){
+    g <- sf::st_as_sf(g)
   }
+  if(!is.na(sr)){
+    net <- sf::st_transform(net, sr)
+    g <- sf::st_transform(g, sr)
+  }
+
+  if (type == "lines" ) {
+    ncolnet <- ncol(sf::st_set_geometry(net, NULL))
+    namesnet <- names(sf::st_set_geometry(net, NULL))
+    net$LKM <- sf::st_length(sf::st_cast(net[sf::st_dimension(net) == 1,]))
+    netg <- suppressWarnings(st_intersection(net, g))
+    netg$LKM2 <- sf::st_length(sf::st_cast(netg[sf::st_dimension(netg) == 1,]))
+    xgg <- data.table::data.table(netg)
+    xgg[, 1:ncolnet] <- xgg[, 1:ncolnet] * as.numeric(xgg$LKM2/xgg$LKM)
+    dfm <- xgg[, lapply(.SD, sum, na.rm=TRUE),
+               by = id,
+               .SDcols = namesnet]
+    names(dfm) <- c("id", namesnet)
+    gx <- data.frame(id = g$id)
+    gx <- merge(gx, dfm, by="id", all.x = T)
+    gx <- sf::st_sf(gx, geometry = g$geometry)
+  # if(array){
+  #     return(GriddedEmissionsArray(gx, rows = , cols, times = ))
+  #   } else{
+      return(gx)
+    # }
+  } else if ( type == "points" ){
+    xgg <- data.table::data.table(sf::st_set_geometry(sf::st_intersection(df, g), NULL))
+    dfm <- xgg[, lapply(.SD, sum, na.rm=TRUE),
+               by = id,
+               .SDcols = namesnet ]
+    names(dfm) <- c("id", namesnet)
+    gx <- data.frame(id = g$id)
+    gx <- merge(gx, dfm, by="id", all.x = T)
+    gx <- sf::st_sf(gx, geometry = g$geometry)
+    return(gx)
+  }
+}